requirements.txt
fastapi
uvicorn
numpy
torch
faiss-cpu
pandas
fasttext
gensim
